requirements.txt
Cython
PyPDF2
beautifulsoup4
faiss-cpu
flask
langchain
langchain-community
langchain-openai
markdown
nltk
numpy
python-docx
rank-bm25
scikit-learn
sentence-transformers
tiktoken
torch
transformers
